Assignment 3: Views and drawing.
+_+_+_+_+_+_+_+_+_+

I feel pretty good about this, except for one part. For some reason, I couldn't get the polygon IBOutlet on PolygonView to properly instantiate with the same polygon as the controller. I set it up in IB, and no matter what I tried, I got two different polygons. So, in -awakeFromNib in the controller, I set the property directly -- which I shouldn't have to do!